+++ api/src/main/kotlin/uy/kohesive/injekt/api/Exceptions.kt
@@ -1,5 +1,5 @@
 package uy.kohesive.injekt.api
 
-public class InjektionException(msg: String) : RuntimeException(msg)
+class InjektionException(msg: String) : RuntimeException(msg)