+++ api/src/main/kotlin/uy/kohesive/injekt/api/TypeInfo.kt
@@ -2,8 +2,7 @@ package uy.kohesive.injekt.api
 
 import java.lang.reflect.*
 
-@Suppress("UNCHECKED_CAST")
-public fun Type.erasedType(): Class<Any> {
+@Suppress("UNCHECKED_CAST") fun Type.erasedType(): Class<Any> {
     return when (this) {
         is Class<*> -> this as Class<Any>
         is ParameterizedType -> this.getRawType().erasedType()
@@ -24,15 +23,15 @@ public fun Type.erasedType(): Class<Any> {
     }
 }
 
-public inline fun <reified T: Any> typeRef(): FullTypeReference<T> = object:FullTypeReference<T>(){}
-public inline fun <reified T: Any> fullType(): FullTypeReference<T> = object:FullTypeReference<T>(){}
+inline fun <reified T: Any> typeRef(): FullTypeReference<T> = object:FullTypeReference<T>(){}
+inline fun <reified T: Any> fullType(): FullTypeReference<T> = object:FullTypeReference<T>(){}
 
-public interface TypeReference<T> {
-    public val type: Type
+interface TypeReference<T> {
+    val type: Type
 }
 
-public abstract class FullTypeReference<T> protected constructor() : TypeReference<T> {
-    override public val type: Type = javaClass.getGenericSuperclass().let { superClass ->
+abstract class FullTypeReference<T> protected constructor() : TypeReference<T> {
+    override val type: Type = javaClass.getGenericSuperclass().let { superClass ->
         if (superClass is Class<*>) {
             throw IllegalArgumentException("Internal error: TypeReference constructed without actual type information")
         }